requirements.txt
apngasm_python~=1.2.1
pyav~=11.0.1
beautifulsoup4~=4.12.2
rookiepy~=0.2.7
imageio~=2.31.5
memory-tempfile~=2.2.3
numpy~=1.26.0
Pillow~=10.0.1
pyoxipng~=8.0.1
python-telegram-bot~=20.6
requests~=2.31.0
rlottie_python~=1.2.1
selenium~=4.13.0
signalstickers-client~=3.3.0
tqdm~=4.66.1
ttkbootstrap-fork-laggykiller~=1.5.1
webp~=0.3.0